lista funcionando
