+++ src/main/java/acme/features/administrator/spam/AdministratorSpamController.java
@@ -3,13 +3,16 @@ package acme.features.administrator.spam;
 import javax.annotation.PostConstruct;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.RequestMapping;
 
 import acme.entities.spam.Spam;
+import acme.framework.components.BasicCommand;
 import acme.framework.controllers.AbstractController;
 import acme.framework.entities.Administrator;
 
-//@Controller
-//@RequestMapping("administrator/spam/")
+@Controller
+@RequestMapping("administrator/spam/")
 public class AdministratorSpamController extends AbstractController<Administrator, Spam>{
 	
 	@Autowired
@@ -18,7 +21,8 @@ public class AdministratorSpamController extends AbstractController<Administrato
 	
 	@PostConstruct
 	protected void initialise() {
-		//super.addBasicCommand(BasicCommand.LIST, this.listService);
+		
+		super.addBasicCommand(BasicCommand.LIST, this.listService);
 	}
 
 }
