+++ src/main/java/acme/features/administrator/spam/AdministratorSpamListService.java
@@ -1,19 +1,18 @@
 package acme.features.administrator.spam;
 
 import java.util.Collection;
-import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import acme.entities.spam.SpamWord;
+import acme.entities.spam.Spam;
 import acme.framework.components.Model;
 import acme.framework.components.Request;
 import acme.framework.entities.Administrator;
 import acme.framework.services.AbstractListService;
 
 @Service
-public class AdministratorSpamListService implements AbstractListService<Administrator, SpamWord>{
+public class AdministratorSpamListService implements AbstractListService<Administrator, Spam>{
 
 	
 	// Internal state ---------------------------------------------------------
@@ -21,45 +20,27 @@ public class AdministratorSpamListService implements AbstractListService<Adminis
 		@Autowired
 		protected AdministratorSpamRepository repository;
 		
-		
+
 	@Override
-	public boolean authorise(final Request<SpamWord> request) {
+	public boolean authorise(final Request<Spam> request) {
 		assert request != null;
 		return true;
 	}
-	
+
 	@Override
-	public void unbind(final Request<SpamWord> request, final SpamWord entity, final Model model) {
+	public void unbind(final Request<Spam> request, final Spam entity, final Model model) {
 		assert request != null;
 		assert entity != null;
 		assert model != null;
 		
-		request.unbind(entity, model, "word");
+		request.unbind(entity, model, "umbral");
+		
 	}
-	
+
 	@Override
-	public Collection<SpamWord> findMany(final Request<SpamWord> request) {
-		assert request  != null;
-		//final List<SpamWord> result;
-		//result = this.repository.findAllSpamWordObjects();
-		return this.repository.findAllSpamWordObjects();
-	}
+	public Collection<Spam> findMany(final Request<Spam> request) {
 
-
-	public List<String> findAllSpamWord(){
-		List<String> result;
-		
-		result=this.repository.findAllSpamWord();
-		
-		return result;
-	}
-	
-	public double umbral() {
-		double result;
-		
-		result=this.repository.umbral();
-		
-		return result;
+		return this.repository.findAllSpam();
 	}
 
 }
