+++ src/main/java/acme/features/administrator/spam/AdministratorSpamRepository.java
@@ -22,7 +22,7 @@ public interface AdministratorSpamRepository extends AbstractRepository {
 	@Query("select s.umbral from Spam s")
 	double umbral();
 	
-	@Query("SELECT spam FROM Spam spam JOIN FETCH spam.spamWords")
+	@Query("select spam from Spam spam JOIN FETCH spam.spamWords")
 	List<Spam> findAllSpam();
 	
 }
