+++ src/main/java/acme/features/manager/task/ManagerTaskUpdateService.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Service;
 
 import acme.entities.roles.Manager;
 import acme.entities.tasks.Task;
-import acme.features.administrator.spam.AdministratorSpamListService;
+import acme.features.administrator.spam.AdministratorSpamWordListService;
 import acme.features.authenticated.task.AuthenticatedTaskRepository;
 import acme.framework.components.Errors;
 import acme.framework.components.Model;
@@ -25,7 +25,7 @@ public class ManagerTaskUpdateService implements AbstractUpdateService<Manager, 
 	// Other Services----------------
 	
 	@Autowired
-	protected AdministratorSpamListService spamService;
+	protected AdministratorSpamWordListService spamService;
 
 @Override
 public boolean authorise(final Request<Task> request) {
