+++ src/main/webapp/WEB-INF/views/administrator/spam-word/i18n_en.messages
@@ -8,8 +8,8 @@
 # purposes.  The copyright owner does not offer any warranties or representations, nor do
 # they accept any liabilities with respect to them.
 
-administrator.spam.list.title.show = Spam
+administrator.spam-word.list.title.show = Spam
 
 
-administrator.spam.list.label.spamWords = Words
-administrator.spam.list.title.spamListing = List of Words Recognised as Spam
+administrator.spam-word.list.label.spamWords = Words
+administrator.spam-word.list.title.spamListing = List of Words Recognised as Spam
