+++ src/main/webapp/WEB-INF/views/administrator/spam-word/i18n_es.messages
@@ -1,4 +1,4 @@
-# i18n_en.messages
+# i18n_es.messages
 #
 # Copyright (C) 2012-2021 Rafael Corchuelo.
 #
@@ -10,6 +10,5 @@
 
 administrator.spam.list.title.show = Spam
 
-
-administrator.spam.list.label.spamWords = Words
-administrator.spam.list.title.spamListing = List of Words Recognised as Spam
+administrator.spam.list.label.spamWords = Palabras
+administrator.spam.list.title.spamListing = Listado de Palabras Reconocidas como Spam
