+++ src/main/webapp/WEB-INF/views/administrator/spam-word/list.jsp
@@ -5,7 +5,7 @@
 
 
 <acme:list>
-			<acme:list-column code="administrator.spam.list.label.spamWords" path="umbral" width="60%"/>
+			<acme:list-column code="administrator.spam-word.list.label.spamWords" path="word" width="60%"/>
 </acme:list>
